requirements.txt
cffi
click
colorama
cryptography
dnspython
Flask
Flask-Cors
itsdangerous
Jinja2
MarkupSafe
numpy
pandas
pycparser
pymongo
python-dateutil
python-dotenv
pytz
six
tzdata
tzlocal
Werkzeug
XlsxWriter
Gunicorn
PyJWT
flask_jwt_extended
xlsxwriter
Werkzeug
flask_mail
apscheduler
msal
flask_socketio
